int error

+++ friend_tracker.py
@@ -90,7 +90,7 @@ def update_data(users_to_update, message):
 # converts datetime to a timedelta string
 def time_to_timedelta(td_int):
 
-    datetime_str = str(datetime.datetime.fromtimestamp(int(int(td_int) / 1e3)))
+    datetime_str = str(datetime.datetime.fromtimestamp(td_int / 1e3))
     format = "%Y-%m-%d %H:%M:%S.%f"
     datetime_obj = datetime.datetime.strptime(datetime_str, format)
